requirements.txt
torch
gym-retro
gym==0.21.0
opencv-python
numpy